test(ui): isolate model metadata lifecycle coverage
